movegen: use buffer instead of array for move generation
Most of the moves generated are often discarded and is mainly used as a
temporary buffer for finding desired moves, therefore it makes more
sense to use a stack allocated buffer instead of dynamically allocating
and freeing one each time. Most of the time the buffer len is 1-2, too
small for dynamic arrays.