move: bit pack "struct move"
Likely not a gain in this project itself but it is a good idea to
bitpack move for future projects I make which will build off move.
Move encoding is typically bitpacked for things like transposition
tables and such and is commonly bitpacked regardless for engines.

This also adds initial work for promoted pieces. On a side note, it is
easier to get the promoted piece from the pgn and store captured piece
instead, but that ties movegen with pgn which makes it less general.
Meanwhile board is already tied to movegen so that is fine in regards to
depending on board state for captures.