pgn: use stretchy buffer for arrays
From skeeto's growable-buf.

This has the advantage of using the natural [] operator to index arrays.
As the reference is not stable, no functions is exposed for modifying
the array, which makes sense as it is read-only. Other array functions are
purposely not exposed to make the elements feel like a regular array.